+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -14,8 +14,6 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import kitchenpos.product.domain.Product;
-
 @Entity
 @Table(name = "menu_product")
 public class MenuProduct {
