+++ src/main/java/kitchenpos/menu/domain/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
