+++ src/main/java/kitchenpos/menu/dto/ProductRequest.java
@@ -1,9 +1,9 @@
-package kitchenpos.product.dto;
+package kitchenpos.menu.dto;
 
 import java.math.BigDecimal;
 
 import kitchenpos.menu.domain.Price;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 public class ProductRequest {
