+++ src/test/java/kitchenpos/menu/MenuProductTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.Price;
 import kitchenpos.menu.domain.Quantity;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 @DisplayName("메뉴 상품 도메인 테스트")
 public class MenuProductTest {
