+++ src/test/java/kitchenpos/menu/MenuProductsTest.java
@@ -13,7 +13,7 @@ import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menu.domain.Price;
 import kitchenpos.menu.domain.Quantity;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 @DisplayName("메뉴 상품들 도메인 테스트")
 public class MenuProductsTest {
