+++ src/test/java/kitchenpos/order/OrderLineItemsTest.java
@@ -19,7 +19,7 @@ import kitchenpos.menu.domain.Quantity;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderLineItems;
 import kitchenpos.order.exception.OrderException;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 @DisplayName("주문항목들 도메인 테스트")
 public class OrderLineItemsTest {
