+++ src/test/java/kitchenpos/order/OrderTest.java
@@ -22,7 +22,7 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.exception.OrderException;
 import kitchenpos.table.domain.NumberOfGuests;
 import kitchenpos.table.domain.OrderTable;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.Product;
 
 @DisplayName("주문 도메인 테스트")
 public class OrderTest {
